requirements.txt
# Core dependencies
torch>=2.0.0
torch_geometric>=2.4.0
numpy>=1.24.0
networkx>=3.0
requests>=2.28.0
PyYAML>=6.0
gymnasium>=0.28.0
